LayoutTests/platform/gtk/media/video-controls-no-scripting-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'video-controls-no-scripting-iframe.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: line 13: Blocked script execution in 'video-controls-no-scripting-iframe.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.

Tests that the built-in controls are always enabled when JavaScript is disabled.

EXPECTED (internals.shadowRoot(video) != 'null') OK
EXPECTED (controls.offsetTop + controls.offsetHeight <= 'video.offsetHeight') OK
EXPECTED (getComputedStyle(controls).opacity == '1') OK

END OF TEST
